requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=mkl
bottleneck=1.3.5=py310h4e76f89_0
bzip2=1.0.8=h1de35cc_0
ca-certificates=2023.05.30=hecd8cb5_0
intel-openmp=2023.1.0=ha357a0b_43547
joblib=1.2.0=py310hecd8cb5_0
libcxx=14.0.6=h9765a3e_0
libffi=3.4.4=hecd8cb5_0
libgfortran=5.0.0=11_3_0_hecd8cb5_28
libgfortran5=11.3.0=h9dfd629_28
llvm-openmp=14.0.6=h0dcd299_0
mkl=2023.1.0=h8e150cf_43559
mkl-service=2.4.0=py310h6c40b1e_1
mkl_fft=1.3.6=py310h3ea8b11_1
mkl_random=1.2.2=py310h3ea8b11_1
ncurses=6.4=hcec6c5f_0
numexpr=2.8.4=py310h827a554_1
numpy=1.25.2=py310h827a554_0
numpy-base=1.25.2=py310ha186be2_0
openssl=3.0.10=hca72f7f_0
pandas=1.5.3=py310h3ea8b11_0
path=16.2.0=pyhd3eb1b0_0
pickle5=0.0.11=pypi_0
pip=23.2.1=py310hecd8cb5_0
python=3.10.12=h5ee71fb_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2022.7=py310hecd8cb5_0
readline=8.2=hca72f7f_0
scikit-learn=1.2.2=py310hcec6c5f_0
scipy=1.11.1=py310hdb2ea58_0
setuptools=68.0.0=py310hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h6c40b1e_0
tbb=2021.8.0=ha357a0b_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h5d9f67b_0
tzdata=2023c=h04d1e81_0
wheel=0.38.4=py310hecd8cb5_0
xz=5.4.2=h6c40b1e_0
zlib=1.2.13=h4dc903c_0
